+++ Lib/v3d/BufferObject.h
@@ -1,40 +1,36 @@
 #ifndef _V3DLIB_V3D_BUFFEROBJECT_H_
 #define _V3DLIB_V3D_BUFFEROBJECT_H_
-//#include <cassert>
 #include <vector>
-//#include <stdint.h>
-//#include "debug.h"
 #include "Common/BufferObject.h"
 
 
 namespace V3DLib {
 namespace v3d {
 
-
 /**
  * This behaves like an array of uint32_t.
  */
 class BufferObject : public V3DLib::BufferObject {
 public:
-	BufferObject(uint32_t size) { alloc_mem(size); }
-	BufferObject() {} 
-	~BufferObject(); 
+  BufferObject(uint32_t size) { alloc_mem(size); }
+  BufferObject() {} 
+  ~BufferObject(); 
 
   uint32_t getHandle() const override { return  (uint32_t) handle; }
 
   void alloc_mem(uint32_t size_in_bytes);
 
-	// Debug methods
-	void fill(uint32_t value);
-	void find_value(uint32_t in_val);
-	void detect_used_blocks();
+  // Debug methods
+  void fill(uint32_t value);
+  void find_value(uint32_t in_val);
+  void detect_used_blocks();
 
 private:
   uint32_t handle  = 0;
 
-	void dealloc_mem();
-	uint32_t &operator[] (int i);
-	uint32_t size_word() const { return size()/sizeof(uint32_t); }  // Returns size in words
+  void dealloc_mem();
+  uint32_t &operator[] (int i);
+  uint32_t size_word() const { return (uint32_t) (size()/sizeof(uint32_t)); }  // Returns size in words
 };
